+++ internal/mcp/json.go
@@ -17,7 +17,6 @@ package mcp
 import (
 	"encoding/json"
 	"fmt"
-	"io"
 
 	"github.com/FerretDB/wire/wirebson"
 	"go.mongodb.org/mongo-driver/v2/bson"
@@ -25,20 +24,70 @@ import (
 	"github.com/FerretDB/FerretDB/v2/internal/util/lazyerrors"
 )
 
-// marshalJSON encodes wirebson.RawDocument into extended JSON.
-func marshalJSON(raw wirebson.RawDocument, jsonDst io.Writer) error {
-	encoder := bson.NewEncoder(bson.NewExtJSONValueWriter(jsonDst, false, false))
+// prepareDocument creates a new bson document from the given pairs of
+// field names and values, which can be used as handler command msg.
+//
+// If any of pair values is nil it's ignored.
+func prepareDocument(pairs ...any) (*wirebson.Document, error) {
+	l := len(pairs)
 
-	if err := encoder.Encode(bson.Raw(raw)); err != nil {
-		return lazyerrors.Error(err)
+	if l%2 != 0 {
+		return nil, lazyerrors.Errorf("invalid number of arguments: %d", l)
 	}
 
-	return nil
+	docPairs := make([]any, 0, l)
+
+	for i := 0; i < l; i += 2 {
+		var err error
+
+		key := pairs[i]
+		v := pairs[i+1]
+
+		switch val := v.(type) {
+		// json.RawMessage is the non-pointer exception.
+		// Other non-pointer types don't need special handling.
+		case json.RawMessage:
+			v, err = unmarshalExtJSON(&val)
+			if err != nil {
+				return nil, err
+			}
+
+		case *json.RawMessage:
+			if val == nil {
+				continue
+			}
+
+			v, err = unmarshalExtJSON(val)
+			if err != nil {
+				return nil, err
+			}
+		case *float32:
+			if val == nil {
+				continue
+			}
+
+			v = float64(*val)
+		case *bool:
+			if val == nil {
+				continue
+			}
+
+			v = *val
+		}
+
+		if v == nil {
+			continue
+		}
+
+		docPairs = append(docPairs, key, v)
+	}
+
+	return wirebson.NewDocument(docPairs...)
 }
 
 // unmarshalSingleJSON takes extended JSON object and unmarshals it into the wirebson composite.
 // If provided json is nil it also returns nil.
-func unmarshalSingleJSON(json *json.RawMessage) (out any, err error) {
+func unmarshalExtJSON(json *json.RawMessage) (out any, err error) {
 	if json == nil {
 		return nil, nil
 	}
